Update WebhookTrigger mutation
