+++ saleor/graphql/webhook/tests/fixtures.py
@@ -1,0 +1,424 @@
+import pytest
+
+from ....webhook.event_types import WebhookEventAsyncType
+
+
+@pytest.fixture
+def async_subscription_webhooks_with_root_objects(
+    subscription_address_created_webhook,
+    subscription_address_updated_webhook,
+    subscription_address_deleted_webhook,
+    subscription_app_installed_webhook,
+    subscription_app_updated_webhook,
+    subscription_app_deleted_webhook,
+    subscription_app_status_changed_webhook,
+    subscription_attribute_created_webhook,
+    subscription_attribute_updated_webhook,
+    subscription_attribute_deleted_webhook,
+    subscription_attribute_value_created_webhook,
+    subscription_attribute_value_updated_webhook,
+    subscription_attribute_value_deleted_webhook,
+    subscription_category_created_webhook,
+    subscription_category_updated_webhook,
+    subscription_category_deleted_webhook,
+    subscription_channel_created_webhook,
+    subscription_channel_updated_webhook,
+    subscription_channel_deleted_webhook,
+    subscription_channel_status_changed_webhook,
+    subscription_gift_card_created_webhook,
+    subscription_gift_card_updated_webhook,
+    subscription_gift_card_deleted_webhook,
+    subscription_gift_card_sent_webhook,
+    subscription_gift_card_status_changed_webhook,
+    subscription_gift_card_metadata_updated_webhook,
+    subscription_menu_created_webhook,
+    subscription_menu_updated_webhook,
+    subscription_menu_deleted_webhook,
+    subscription_menu_item_created_webhook,
+    subscription_menu_item_updated_webhook,
+    subscription_menu_item_deleted_webhook,
+    subscription_shipping_price_created_webhook,
+    subscription_shipping_price_updated_webhook,
+    subscription_shipping_price_deleted_webhook,
+    subscription_shipping_zone_created_webhook,
+    subscription_shipping_zone_updated_webhook,
+    subscription_shipping_zone_deleted_webhook,
+    subscription_shipping_zone_metadata_updated_webhook,
+    subscription_product_updated_webhook,
+    subscription_product_created_webhook,
+    subscription_product_deleted_webhook,
+    subscription_product_media_updated_webhook,
+    subscription_product_media_created_webhook,
+    subscription_product_media_deleted_webhook,
+    subscription_product_metadata_updated_webhook,
+    subscription_product_variant_created_webhook,
+    subscription_product_variant_updated_webhook,
+    subscription_product_variant_deleted_webhook,
+    subscription_product_variant_metadata_updated_webhook,
+    subscription_product_variant_out_of_stock_webhook,
+    subscription_product_variant_back_in_stock_webhook,
+    subscription_order_created_webhook,
+    subscription_order_updated_webhook,
+    subscription_order_confirmed_webhook,
+    subscription_order_fully_paid_webhook,
+    subscription_order_refunded_webhook,
+    subscription_order_fully_refunded_webhook,
+    subscription_order_paid_webhook,
+    subscription_order_cancelled_webhook,
+    subscription_order_expired_webhook,
+    subscription_order_fulfilled_webhook,
+    subscription_order_metadata_updated_webhook,
+    subscription_order_bulk_created_webhook,
+    subscription_draft_order_created_webhook,
+    subscription_draft_order_updated_webhook,
+    subscription_draft_order_deleted_webhook,
+    subscription_promotion_created_webhook,
+    subscription_promotion_updated_webhook,
+    subscription_promotion_deleted_webhook,
+    subscription_promotion_started_webhook,
+    subscription_promotion_ended_webhook,
+    subscription_sale_created_webhook,
+    subscription_sale_updated_webhook,
+    subscription_sale_deleted_webhook,
+    subscription_sale_toggle_webhook,
+    subscription_invoice_requested_webhook,
+    subscription_invoice_deleted_webhook,
+    subscription_invoice_sent_webhook,
+    subscription_fulfillment_created_webhook,
+    subscription_fulfillment_canceled_webhook,
+    subscription_fulfillment_approved_webhook,
+    subscription_fulfillment_metadata_updated_webhook,
+    subscription_customer_created_webhook,
+    subscription_customer_updated_webhook,
+    subscription_customer_deleted_webhook,
+    subscription_customer_metadata_updated_webhook,
+    subscription_collection_created_webhook,
+    subscription_collection_updated_webhook,
+    subscription_collection_deleted_webhook,
+    subscription_collection_metadata_updated_webhook,
+    subscription_checkout_created_webhook,
+    subscription_checkout_updated_webhook,
+    subscription_checkout_fully_paid_webhook,
+    subscription_checkout_metadata_updated_webhook,
+    subscription_page_created_webhook,
+    subscription_page_updated_webhook,
+    subscription_page_deleted_webhook,
+    subscription_page_type_created_webhook,
+    subscription_page_type_updated_webhook,
+    subscription_page_type_deleted_webhook,
+    subscription_permission_group_created_webhook,
+    subscription_permission_group_updated_webhook,
+    subscription_permission_group_deleted_webhook,
+    subscription_product_created_multiple_events_webhook,
+    subscription_staff_created_webhook,
+    subscription_staff_updated_webhook,
+    subscription_staff_deleted_webhook,
+    subscription_transaction_item_metadata_updated_webhook,
+    subscription_translation_created_webhook,
+    subscription_translation_updated_webhook,
+    subscription_warehouse_created_webhook,
+    subscription_warehouse_updated_webhook,
+    subscription_warehouse_deleted_webhook,
+    subscription_warehouse_metadata_updated_webhook,
+    subscription_voucher_created_webhook,
+    subscription_voucher_updated_webhook,
+    subscription_voucher_deleted_webhook,
+    subscription_voucher_webhook_with_meta,
+    subscription_voucher_metadata_updated_webhook,
+    address,
+    app,
+    numeric_attribute,
+    category,
+    channel_PLN,
+    gift_card,
+    menu_item,
+    shipping_method,
+    product,
+    fulfilled_order,
+    promotion,
+    promotion_converted_from_sale,
+    fulfillment,
+    stock,
+    customer_user,
+    collection,
+    checkout,
+    page,
+    permission_group_manage_users,
+    shipping_zone,
+    staff_user,
+    voucher,
+    warehouse,
+    translated_attribute,
+    transaction_item_created_by_app,
+    product_media_image,
+):
+    events = WebhookEventAsyncType
+    attr = numeric_attribute
+    attr_value = attr.values.first()
+    menu = menu_item.menu
+    order = fulfilled_order
+    invoice = order.invoices.first()
+    page_type = page.page_type
+    transaction_item_created_by_app.use_old_id = True
+    transaction_item_created_by_app.save()
+
+    return {
+        events.ADDRESS_UPDATED: [subscription_address_updated_webhook, address],
+        events.ADDRESS_CREATED: [subscription_address_created_webhook, address],
+        events.ADDRESS_DELETED: [subscription_address_deleted_webhook, address],
+        events.APP_UPDATED: [subscription_app_updated_webhook, app],
+        events.APP_DELETED: [subscription_app_deleted_webhook, app],
+        events.APP_INSTALLED: [subscription_app_installed_webhook, app],
+        events.APP_STATUS_CHANGED: [subscription_app_status_changed_webhook, app],
+        events.ATTRIBUTE_CREATED: [subscription_attribute_created_webhook, attr],
+        events.ATTRIBUTE_UPDATED: [subscription_attribute_updated_webhook, attr],
+        events.ATTRIBUTE_DELETED: [subscription_attribute_deleted_webhook, attr],
+        events.ATTRIBUTE_VALUE_UPDATED: [
+            subscription_attribute_value_updated_webhook,
+            attr_value,
+        ],
+        events.ATTRIBUTE_VALUE_CREATED: [
+            subscription_attribute_value_created_webhook,
+            attr_value,
+        ],
+        events.ATTRIBUTE_VALUE_DELETED: [
+            subscription_attribute_value_deleted_webhook,
+            attr_value,
+        ],
+        events.CATEGORY_CREATED: [subscription_category_created_webhook, category],
+        events.CATEGORY_UPDATED: [subscription_category_updated_webhook, category],
+        events.CATEGORY_DELETED: [subscription_category_deleted_webhook, category],
+        events.CHANNEL_CREATED: [subscription_channel_created_webhook, channel_PLN],
+        events.CHANNEL_UPDATED: [subscription_channel_updated_webhook, channel_PLN],
+        events.CHANNEL_DELETED: [subscription_channel_deleted_webhook, channel_PLN],
+        events.CHANNEL_STATUS_CHANGED: [
+            subscription_channel_status_changed_webhook,
+            channel_PLN,
+        ],
+        events.GIFT_CARD_CREATED: [subscription_gift_card_created_webhook, gift_card],
+        events.GIFT_CARD_UPDATED: [subscription_gift_card_updated_webhook, gift_card],
+        events.GIFT_CARD_DELETED: [subscription_gift_card_deleted_webhook, gift_card],
+        events.GIFT_CARD_SENT: [subscription_gift_card_sent_webhook, gift_card],
+        events.GIFT_CARD_STATUS_CHANGED: [
+            subscription_gift_card_status_changed_webhook,
+            gift_card,
+        ],
+        events.GIFT_CARD_METADATA_UPDATED: [
+            subscription_gift_card_metadata_updated_webhook,
+            gift_card,
+        ],
+        events.MENU_CREATED: [subscription_menu_created_webhook, menu],
+        events.MENU_UPDATED: [subscription_menu_updated_webhook, menu],
+        events.MENU_DELETED: [subscription_menu_deleted_webhook, menu],
+        events.MENU_ITEM_CREATED: [subscription_menu_item_created_webhook, menu_item],
+        events.MENU_ITEM_UPDATED: [subscription_menu_item_updated_webhook, menu_item],
+        events.MENU_ITEM_DELETED: [subscription_menu_item_deleted_webhook, menu_item],
+        events.ORDER_CREATED: [subscription_order_created_webhook, order],
+        events.ORDER_UPDATED: [subscription_order_updated_webhook, order],
+        events.ORDER_CONFIRMED: [subscription_order_confirmed_webhook, order],
+        events.ORDER_FULLY_PAID: [subscription_order_fully_paid_webhook, order],
+        events.ORDER_PAID: [subscription_order_paid_webhook, order],
+        events.ORDER_REFUNDED: [subscription_order_refunded_webhook, order],
+        events.ORDER_FULLY_REFUNDED: [subscription_order_fully_refunded_webhook, order],
+        events.ORDER_FULFILLED: [subscription_order_fulfilled_webhook, order],
+        events.ORDER_CANCELLED: [subscription_order_cancelled_webhook, order],
+        events.ORDER_EXPIRED: [subscription_order_expired_webhook, order],
+        events.ORDER_METADATA_UPDATED: [
+            subscription_order_metadata_updated_webhook,
+            order,
+        ],
+        events.ORDER_BULK_CREATED: [subscription_order_bulk_created_webhook, order],
+        events.DRAFT_ORDER_CREATED: [subscription_draft_order_created_webhook, order],
+        events.DRAFT_ORDER_UPDATED: [subscription_draft_order_updated_webhook, order],
+        events.DRAFT_ORDER_DELETED: [subscription_draft_order_deleted_webhook, order],
+        events.PRODUCT_CREATED: [subscription_product_created_webhook, product],
+        events.PRODUCT_UPDATED: [subscription_product_updated_webhook, product],
+        events.PRODUCT_DELETED: [subscription_product_deleted_webhook, product],
+        events.PRODUCT_MEDIA_CREATED: [
+            subscription_product_media_created_webhook,
+            product_media_image,
+        ],
+        events.PRODUCT_MEDIA_UPDATED: [
+            subscription_product_media_updated_webhook,
+            product_media_image,
+        ],
+        events.PRODUCT_MEDIA_DELETED: [
+            subscription_product_media_deleted_webhook,
+            product_media_image,
+        ],
+        events.PRODUCT_METADATA_UPDATED: [
+            subscription_product_metadata_updated_webhook,
+            product,
+        ],
+        events.PRODUCT_VARIANT_CREATED: [
+            subscription_product_variant_created_webhook,
+            product,
+        ],
+        events.PRODUCT_VARIANT_UPDATED: [
+            subscription_product_variant_updated_webhook,
+            product,
+        ],
+        events.PRODUCT_VARIANT_OUT_OF_STOCK: [
+            subscription_product_variant_out_of_stock_webhook,
+            stock,
+        ],
+        events.PRODUCT_VARIANT_BACK_IN_STOCK: [
+            subscription_product_variant_back_in_stock_webhook,
+            stock,
+        ],
+        events.PRODUCT_VARIANT_DELETED: [
+            subscription_product_variant_deleted_webhook,
+            product,
+        ],
+        events.PRODUCT_VARIANT_METADATA_UPDATED: [
+            subscription_product_variant_metadata_updated_webhook,
+            product,
+        ],
+        events.PROMOTION_CREATED: [subscription_promotion_created_webhook, promotion],
+        events.PROMOTION_UPDATED: [subscription_promotion_updated_webhook, promotion],
+        events.PROMOTION_DELETED: [subscription_promotion_deleted_webhook, promotion],
+        events.PROMOTION_STARTED: [subscription_promotion_started_webhook, promotion],
+        events.PROMOTION_ENDED: [subscription_promotion_ended_webhook, promotion],
+        events.SALE_CREATED: [
+            subscription_sale_created_webhook,
+            promotion_converted_from_sale,
+        ],
+        events.SALE_UPDATED: [
+            subscription_sale_updated_webhook,
+            promotion_converted_from_sale,
+        ],
+        events.SALE_DELETED: [
+            subscription_sale_deleted_webhook,
+            promotion_converted_from_sale,
+        ],
+        events.SALE_TOGGLE: [
+            subscription_sale_toggle_webhook,
+            promotion_converted_from_sale,
+        ],
+        events.INVOICE_REQUESTED: [subscription_invoice_requested_webhook, invoice],
+        events.INVOICE_DELETED: [subscription_invoice_deleted_webhook, invoice],
+        events.INVOICE_SENT: [subscription_invoice_sent_webhook, invoice],
+        events.FULFILLMENT_CREATED: [
+            subscription_fulfillment_created_webhook,
+            fulfillment,
+        ],
+        events.FULFILLMENT_CANCELED: [
+            subscription_fulfillment_canceled_webhook,
+            fulfillment,
+        ],
+        events.FULFILLMENT_APPROVED: [
+            subscription_fulfillment_approved_webhook,
+            fulfillment,
+        ],
+        events.FULFILLMENT_METADATA_UPDATED: [
+            subscription_fulfillment_metadata_updated_webhook,
+            fulfillment,
+        ],
+        events.CUSTOMER_CREATED: [subscription_customer_created_webhook, customer_user],
+        events.CUSTOMER_UPDATED: [subscription_customer_updated_webhook, customer_user],
+        events.CUSTOMER_METADATA_UPDATED: [
+            subscription_customer_metadata_updated_webhook,
+            customer_user,
+        ],
+        events.COLLECTION_CREATED: [
+            subscription_collection_created_webhook,
+            collection,
+        ],
+        events.COLLECTION_UPDATED: [
+            subscription_collection_updated_webhook,
+            collection,
+        ],
+        events.COLLECTION_DELETED: [
+            subscription_collection_deleted_webhook,
+            collection,
+        ],
+        events.COLLECTION_METADATA_UPDATED: [
+            subscription_collection_metadata_updated_webhook,
+            collection,
+        ],
+        events.CHECKOUT_CREATED: [subscription_checkout_created_webhook, checkout],
+        events.CHECKOUT_UPDATED: [subscription_checkout_updated_webhook, checkout],
+        events.CHECKOUT_FULLY_PAID: [
+            subscription_checkout_fully_paid_webhook,
+            checkout,
+        ],
+        events.CHECKOUT_METADATA_UPDATED: [
+            subscription_checkout_metadata_updated_webhook,
+            checkout,
+        ],
+        events.PAGE_CREATED: [subscription_page_created_webhook, page],
+        events.PAGE_UPDATED: [subscription_page_updated_webhook, page],
+        events.PAGE_DELETED: [subscription_page_deleted_webhook, page],
+        events.PAGE_TYPE_CREATED: [subscription_page_type_created_webhook, page_type],
+        events.PAGE_TYPE_UPDATED: [subscription_page_type_updated_webhook, page_type],
+        events.PAGE_TYPE_DELETED: [subscription_page_type_deleted_webhook, page_type],
+        events.PERMISSION_GROUP_CREATED: [
+            subscription_permission_group_created_webhook,
+            permission_group_manage_users,
+        ],
+        events.PERMISSION_GROUP_UPDATED: [
+            subscription_permission_group_updated_webhook,
+            permission_group_manage_users,
+        ],
+        events.PERMISSION_GROUP_DELETED: [
+            subscription_permission_group_deleted_webhook,
+            permission_group_manage_users,
+        ],
+        events.SHIPPING_PRICE_CREATED: [
+            subscription_shipping_price_created_webhook,
+            shipping_method,
+        ],
+        events.SHIPPING_PRICE_UPDATED: [
+            subscription_shipping_price_updated_webhook,
+            shipping_method,
+        ],
+        events.SHIPPING_PRICE_DELETED: [
+            subscription_shipping_price_deleted_webhook,
+            shipping_method,
+        ],
+        events.SHIPPING_ZONE_CREATED: [
+            subscription_shipping_zone_created_webhook,
+            shipping_zone,
+        ],
+        events.SHIPPING_ZONE_UPDATED: [
+            subscription_shipping_zone_updated_webhook,
+            shipping_zone,
+        ],
+        events.SHIPPING_ZONE_DELETED: [
+            subscription_shipping_zone_deleted_webhook,
+            shipping_zone,
+        ],
+        events.SHIPPING_ZONE_METADATA_UPDATED: [
+            subscription_shipping_zone_metadata_updated_webhook,
+            shipping_zone,
+        ],
+        events.STAFF_CREATED: [subscription_staff_created_webhook, staff_user],
+        events.STAFF_UPDATED: [subscription_staff_updated_webhook, staff_user],
+        events.STAFF_DELETED: [subscription_staff_deleted_webhook, staff_user],
+        events.TRANSACTION_ITEM_METADATA_UPDATED: [
+            subscription_transaction_item_metadata_updated_webhook,
+            transaction_item_created_by_app,
+        ],
+        events.TRANSLATION_CREATED: [
+            subscription_translation_created_webhook,
+            translated_attribute,
+        ],
+        events.TRANSLATION_UPDATED: [
+            subscription_translation_updated_webhook,
+            translated_attribute,
+        ],
+        events.VOUCHER_CREATED: [subscription_voucher_created_webhook, voucher],
+        events.VOUCHER_UPDATED: [subscription_voucher_updated_webhook, voucher],
+        events.VOUCHER_DELETED: [subscription_voucher_deleted_webhook, voucher],
+        events.VOUCHER_METADATA_UPDATED: [
+            subscription_voucher_metadata_updated_webhook,
+            voucher,
+        ],
+        events.WAREHOUSE_CREATED: [subscription_warehouse_created_webhook, warehouse],
+        events.WAREHOUSE_UPDATED: [subscription_warehouse_updated_webhook, warehouse],
+        events.WAREHOUSE_DELETED: [subscription_warehouse_deleted_webhook, warehouse],
+        events.WAREHOUSE_METADATA_UPDATED: [
+            subscription_warehouse_metadata_updated_webhook,
+            warehouse,
+        ],
+    }
